Moved Testcase to namespace test class
