+++ hadoop-tools/hadoop-azure/src/main/java/org/apache/hadoop/fs/azurebfs/AbfsConfiguration.java
@@ -1530,7 +1530,7 @@ public class AbfsConfiguration{
 
   @VisibleForTesting
   public void setIsNamespaceEnabledAccount(Trilean isNamespaceEnabledAccount) {
-    isNamespaceEnabled = isNamespaceEnabledAccount;
+    this.isNamespaceEnabled = isNamespaceEnabledAccount;
   }
 
   /**
