+++ hadoop-tools/hadoop-azure/src/main/java/org/apache/hadoop/fs/azurebfs/services/AbfsClient.java
@@ -1716,13 +1716,14 @@ public abstract class AbfsClient implements Closeable {
    * Checks if the namespace is enabled.
    *
    * @return True if the namespace is enabled, false otherwise.
+   * @throws AbfsDriverException if the conversion fails.
    */
-  public boolean getIsNamespaceEnabled() throws TrileanConversionException {
+  public boolean getIsNamespaceEnabled() throws AbfsDriverException {
     try {
       return abfsConfiguration.getIsNamespaceEnabledAccount().toBoolean();
     } catch (TrileanConversionException ex) {
       LOG.error("Failed to convert namespace enabled account property to boolean", ex);
-      throw ex;
+      throw new AbfsDriverException("Failed to determine if namespace is enabled", ex);
     }
   }
